requirements.txt
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-modeltranslation==0.18.11
django-query-counter==0.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
inflection==0.5.1
ipython==8.21.0
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
pyTelegramBotAPI==4.15.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tabulate==0.9.0
telebot==0.0.5
traitlets==5.14.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
wcwidth==0.2.13
